+++ examples/example1.py
@@ -6,9 +6,10 @@ log_file_1 = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__fil
 log_file_2 = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)),'../log/fake_2.log'))
 
 # tail them
-tail.files(
-    debug=True,
-    paths=[
+tail.run(
+    # debug=True,
+    # lines=10
+    sources=[
         log_file_1, 
         log_file_2
     ]